+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [3.1.0] - 2020-11-20
+### Changed
+- Use Symfony 5.1 instead 5.0
+### Removed
+- Classes: `RequestExtbasePersistenceClassesEvent`
+- Methods: `ConfigLoader::loadFromRequestExtbasePersistenceClasses`
+
 ## [3.0.4] - 2020-07-16
 ### Fixed
 - Initialize TSFE for symfony routes
@@ -254,7 +262,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Release for TYPO3 8.7 LTS
 
-[Unreleased]: https://github.com/Bartacus/BartacusBundle/compare/3.0.4...HEAD
+[Unreleased]: https://github.com/Bartacus/BartacusBundle/compare/3.1.0...HEAD
+[3.1.0]: https://github.com/Bartacus/BartacusBundle/compare/3.0.4...3.1.0
 [3.0.4]: https://github.com/Bartacus/BartacusBundle/compare/3.0.3...3.0.4
 [3.0.3]: https://github.com/Bartacus/BartacusBundle/compare/3.0.2...3.0.3
 [3.0.2]: https://github.com/Bartacus/BartacusBundle/compare/3.0.1...3.0.2